# Fixes and poppler performance gain