--- a/.gitignore
+++ b/.gitignore
@@ -1,0 +1,1 @@
+test.png