--- a/pdfdraw_poppler.go
+++ b/pdfdraw_poppler.go
@@ -6,24 +6,8 @@ package pdfdraw
 
 /*
 #cgo pkg-config: poppler-glib cairo
-#include <stdlib.h>
 #include <poppler.h>
 #include <cairo.h>
-
-static char *path_to_uri(char *path) {
-	GError *err = NULL;
-	gchar *absfn = NULL;
-
-	if (g_path_is_absolute(path))
-		absfn = g_strdup(path);
-	else {
-		gchar *tmp = g_get_current_dir();
-		absfn = g_build_filename(tmp, path, NULL);
-		free(tmp);
-	}
-
-	return (char *) g_filename_to_uri(absfn, NULL, &err);
-}
 */
 import "C"
 
@@ -31,13 +15,15 @@ import (
 	"errors"
 	"image"
 	"image/color"
+	"io/ioutil"
 	"reflect"
 
 	"unsafe"
 )
 
 type popplerDocument struct {
-	doc *C.PopplerDocument
+	doc  *C.PopplerDocument
+	data []byte // keep for gc
 }
 
 type popplerPage struct {
@@ -48,19 +34,16 @@ func init() {
 	RegisterBackend("poppler", popplerOpenDoc)
 }
 
-func popplerOpenDoc(path string) (doc Document, err error) {
-	uri := C.path_to_uri(C.CString(path))
-	if uri == nil {
-		return nil, errors.New("unable to convert path to uri")
+func popplerOpenDoc(path string) (Document, error) {
+	data, err := ioutil.ReadFile(path)
+	if err != nil {
+		return nil, err
 	}
-	defer C.free(unsafe.Pointer(uri))
-
-	pd := new(popplerDocument)
-	pd.doc = C.poppler_document_new_from_file(uri, nil, nil)
-	if pd == nil {
+	doc := C.poppler_document_new_from_data((*C.char)(unsafe.Pointer(&data[0])), C.int(len(data)), nil, nil)
+	if doc == nil {
 		return nil, errors.New("unable to open file")
 	}
-	return pd, nil
+	return &popplerDocument{doc, data}, nil
 }
 
 func (doc *popplerDocument) Close() error {