Hi.  Thanks for the lib.  Very handy for my project.  Fixed a few deprecation warnings, and a segfault (now requires closing doc when done), and boosted the performance quite a bit.  

```
benchmark             old ns/op     new ns/op     delta
BenchmarkRender-4     502454301     6151739       -98.78%

benchmark             old allocs     new allocs     delta
BenchmarkRender-4     4              4              +0.00%

benchmark             old bytes     new bytes     delta
BenchmarkRender-4     5931088       5931092       +0.00%
```